synapse/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
groq==0.9.0
httpx==0.27.0
supabase==2.5.0
google-api-python-client==2.134.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
slack-sdk==3.27.2
tiktoken==0.7.0
pydantic==2.7.3
pydantic-settings==2.3.3
python-multipart==0.0.9
numpy==1.26.4
